requirements.txt
# Core dependencies
transformers==4.52.4
accelerate==0.34.2
deepspeed==0.8.3
flash-attn==2.5.8

# Data & evaluation
datasets>=2.14.0
scikit-learn==1.2.2
scipy==1.10.1
numpy==1.24.4
pandas==2.2.3

# Multimodal & Vision
opencv-python==4.7.0.72
timm==1.0.11
Pillow==9.5.0
einops==0.6.0

# Training utilities
omegaconf==2.3.0
yacs==0.1.8
tqdm==4.64.1
pyyaml==6.0
filelock==3.16.1

# Logging & visualization
tensorboardX==2.6.2.2
matplotlib==3.7.0
seaborn==0.13.0
rich==13.9.4

# Evaluation (optional)
lpips==0.1.4
sentence-transformers==2.2.2
evaluate==0.4.2
